// hvn-telegram-bot/src/shared/types/data.ts
export type User = {
  uid: string;
  email: string;
  displayName: string;
  role: 'admin' | 'employee';
  id: string; // Firestore document ID is the same as uid
  telegramUsername?: string;
};